+++ CHANGELOG.md
@@ -15,8 +15,10 @@
     * Remove `insertIntoLocalizedString` methods
     * Remove translation string static helper methods
   * BTDropInResult
-    * Rename to `BTDropInResult.mostRecentPaymentMethod`
+    * Rename `BTDropInResult.fetch` to `BTDropInResult.mostRecentPaymentMethod`
     * Fix bug where fetching most recent payment methods would return an empty `BTDropInResult` instead of `nil` when the customer doesn't have any vaulted payment methods
+    * Rename `BTUIKPaymentOptionType` enum to `BTDropInPaymentMethodType`
+    * Rename `paymentOptionType` property to `paymentMethodType`
 
 ## 9.0.0-beta1 (2021-03-29)